Updated Qty to decimal

+++ src/DelayedPaymentService/Data/Models/PurchasedProduct.cs
@@ -15,7 +15,7 @@
 
         public string ExternalId { get; set; } = null!;
 
-        public int BoughtQuantity { get; set; }
+        public decimal BoughtQuantity { get; set; }
 
         public decimal PurchasePrice { get; set; }
